requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
xgboost
pyyaml
python-dotenv
requests
psycopg2-binary
